feat.message api 加入 user 判斷

+++ src/module/resource/message.py
@@ -2,7 +2,7 @@ import uuid
 from flask import request
 from flask_restful import Resource
 
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from src.model.messages import Message as MessageModel
 from src.model.messages import Reply as ReplyModel
@@ -21,11 +21,11 @@ class MessageBoard(Resource):
     @jwt_required()
     def post(self):
         new_message = request.get_json()
-        print("new_message =>", new_message)
 
         message = MessageModel(
             message_id=str(uuid.uuid4()).replace("-", ""),
             content=new_message.get("content"),
+            create_account=get_jwt_identity(),
         )
 
         message.add()
@@ -42,8 +42,11 @@ class MessageBoard(Resource):
 
         if message:
             temp_message = request.get_json()
-            message.content = temp_message.get("content")
-            message.update()
+            if not message.create_account == get_jwt_identity():
+                return {"code": "0", "data": None, "message": "拒絕不同帳號更新留言"}
+            else:
+                message.content = temp_message.get("content")
+                message.update()
 
             message_list = [
                 message.to_dict() for message in MessageModel.get_message_list()
@@ -60,7 +63,10 @@ class MessageBoard(Resource):
         message = MessageModel.get_by_message_id(message_id)
 
         if message:
-            message.delete()
+            if not message.create_account == get_jwt_identity():
+                return {"code": "0", "data": None, "message": "拒絕不同帳號刪除留言"}
+            else:
+                message.delete()
 
             message_list = [
                 message.to_dict() for message in MessageModel.get_message_list()
